Fix mixpanel initializing
