+++ website/src/app/layout.tsx
@@ -1,5 +1,10 @@
 import RootLayout from "@/components/RootLayout";
+import Providers from "@/components/Providers";
 
 export default function Layout({ children }: { children: React.ReactNode }) {
-  return <RootLayout>{children}</RootLayout>;
+  return (
+    <Providers>
+      <RootLayout>{children}</RootLayout>
+    </Providers>
+  );
 }
